+++ src/main/java/uk/ac/strath/cs112/Main.java
@@ -1,0 +1,11 @@
+package uk.ac.strath.cs112;
+
+public class Main {
+
+  public static void main(String... args) {
+
+    TaxiCompany taxiCompany = new TaxiCompany(new City());
+
+  }
+
+}
