+++ src/main/java/uk/ac/strath/cs112/MissingPassengerException.java
@@ -1,5 +1,7 @@
 package uk.ac.strath.cs112;
 
+import uk.ac.strath.cs112.model.impl.actor.Vehicle;
+
 /**
  * Indicate that there was no passenger at a pickup point.
  *
